requirements.txt
torch>=1.10.0
psutil>=5.9.0
pyyaml>=6.0
matplotlib>=3.5.0
numpy>=1.22.0